turn off TextInput in Markdown bodies
As discovered in issue #203, when you have a <Text onPress /> nested
within a <TextInput/>, the TextInput eats all touch events (I believe
to handle selection handles and whatnot).

Until this is fixed, probably at the RN level, which means I should
file a bug there, we won't be able to use TextInput to wrap any text
that needs to be tappable.