Return null instead of the searchbar component
In order to fix it, this requires quite the hole
of changes. Porting over the updated interfaces
from AAO would involve an upgrade in both react
and react native due to the adoption of hooks.

This is a reduction in functionality across the
the app (orgs, dictionary, map) but gets us
building until we're ready to take on another big
upgrade.